Add functionality for training programs, get, post, put, delete if program hasn't happened yet

+++ models/training-program-model.js
@@ -1,9 +1,24 @@
 'use strict';
 
 const sqlite3 = require('sqlite3').verbose();
-
+const moment = require('moment');
 //need to require in the database once named
 const db = new sqlite3.Database('./db/bangazonStore.sqlite');
+var DATE_RFC2822 = "ddd, DD MMM YYYY HH:mm:ss ZZ";
+
+
+let get_date = (id) => {
+  return new Promise((resolve,reject)=>{
+    db.get(`SELECT start_date FROM programs WHERE program_id = ${id}`, (err, date)=>{
+      if(err) {
+        console.log(`error from model`, err);
+        return reject (err)
+      }
+      resolve(date) 
+    });
+  });
+};
+
 
 module.exports = {
   get_all: () => {
@@ -41,7 +56,7 @@ module.exports = {
   update_training_model: (id, body) => {
     return new Promise((resolve, reject) => {
       db.run(`DELETE FROM programs WHERE program_id = ${id}`);
-      db.run(`INSERT INTO programs (user_id, first_name, last_name, phone, email, address_street, address_city, address_state, address_zip) VALUES(
+      db.run(`INSERT INTO programs (program_id, start_date, end_date) VALUES(
         ${id}, "${body.start_date}", "${body.end_date}")`,
       (err, data) => {
         if (err) return reject(err);
@@ -52,11 +67,18 @@ module.exports = {
   //TODO:make programs delete based on start date
   delete_one: (id) => {
     return new Promise((resolve, reject) => {
-      db.run(`DELETE FROM programs WHERE program_id = ${id}`, 
-      (err, data) => {
-        if (err) return reject(err);
-        resolve(data);
-      });
+        get_date(id)
+        .then((date)=>{
+          console.log(moment().isBefore(date.start_date))
+          console.log("today", moment(), "date", date.start_date);
+          if(moment().isBefore(date.start_date)===true){
+            db.run(`DELETE FROM programs WHERE program_id = ${id}`, 
+            (err, data) => {
+              if (err) return reject(err);
+              resolve(data);
+            })
+          }
+        });
     });
   }
 }
